+++ lib/presentation/auth/sign_in/user_pass_page.dart
@@ -104,12 +104,12 @@ class _UserPassPageState extends State<UserPassPage> {
               TextFormField(
                 style: appThemeData.textTheme.bodyText2,
                 maxLength: _userPass.maxLength,
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   prefixIcon: Icon(
                     Icons.lock,
                     // color: appThemeData.colorScheme.onPrimary,
                   ),
-                  errorStyle: const TextStyle(
+                  errorStyle: TextStyle(
                     height: 1.1,
                   ),
                   errorMaxLines: 5,
