+++ lib/presentation/auth/sign_in/widgets/phone_number_widget.dart
@@ -44,7 +44,7 @@ class _PhoneNumbetWidgetState extends State<PhoneNumbetWidget> {
             keyboardType: TextInputType.number,
             maxLength: 10,
             decoration: InputDecoration(
-              prefixIcon: Icon(
+              prefixIcon: const Icon(
                 Icons.phone,
                 // color: appThemeData.colorScheme.onPrimary,
               ),
