+++ lib/presentation/user_account/widgets/order_overview_body.dart
@@ -2,12 +2,13 @@ import 'package:flowers_app/assets/texts/app_text.dart';
 import 'package:flowers_app/dev/log/log.dart';
 import 'package:flowers_app/domain/auth/app_user.dart';
 import 'package:flowers_app/domain/order/order.dart';
+import 'package:flowers_app/domain/order/order_header.dart';
 import 'package:flowers_app/domain/order/order_list.dart';
-import 'package:flowers_app/presentation/core/app_theme.dart';
 import 'package:flowers_app/presentation/core/widgets/critical_error_widget.dart';
 import 'package:flowers_app/presentation/core/widgets/in_pogress_overlay.dart';
 import 'package:flowers_app/presentation/purchase/purchase_overview/widgets/error_purchase_card.dart';
 import 'package:flowers_app/presentation/user_account/widgets/order_card.dart';
+import 'package:flowers_app/presentation/user_account/widgets/order_header_card.dart';
 import 'package:flutter/material.dart';
 
 class OrderOverviewBody extends StatelessWidget {
@@ -21,7 +22,7 @@ class OrderOverviewBody extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return StreamBuilder<List<dynamic>>(
+    return StreamBuilder<List<Order>>(
       stream: orderList.dataStream,
       builder: (context, snapshot) {
         return RefreshIndicator(
@@ -35,20 +36,20 @@ class OrderOverviewBody extends StatelessWidget {
 
   Widget _buildListViewWidget(
     BuildContext context, 
-    AsyncSnapshot<List<dynamic>> snapshot,
+    AsyncSnapshot<List<Order>> snapshot,
   ) {
-    final List<dynamic> _orders = snapshot.data ?? List.empty();
+    final _orders = snapshot.data ?? [];
     final List<dynamic> orders = [];
-    String order_purchase_id = '-1';
-    for (var order in _orders) {
-      if ('${order['purchase/id']}' != order_purchase_id) {
-        order_purchase_id = '${order['purchase/id']}';
+    String orderPurchaseId = '-1';
+    for (final order in _orders) {
+      if ('${order['purchase/id']}' != orderPurchaseId) {
+        orderPurchaseId = '${order['purchase/id']}';
         orders.add(
-          OrderHeader(order: order as Order),
+          OrderHeader(order: order),
         );
       }
       orders.add(
-        order as Order,
+        order,
       );
     }
     log('[$OrderOverviewBody._buildListView]');
@@ -66,7 +67,9 @@ class OrderOverviewBody extends StatelessWidget {
           itemCount: orders.length,
           itemBuilder: (context, index) {
             if (orders[index] is OrderHeader) {
-              return OrderHeaderCard(orderHeader: orders[index] as OrderHeader);
+              return OrderHeaderCard(
+                orderHeader: orders[index] as OrderHeader,
+              );
             } else {
               final order = orders[index] as Order;
               if (order.valid()) {
@@ -88,50 +91,5 @@ class OrderOverviewBody extends StatelessWidget {
         message: AppText.loading,
       );
     }
-  }
-}
-
-class OrderHeaderCard extends StatelessWidget {
-  final OrderHeader _orderHeader;
-  const OrderHeaderCard({
-    Key? key,
-    required OrderHeader orderHeader,
-  }) : 
-    _orderHeader = orderHeader,
-    super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: appThemeData.colorScheme.secondaryContainer,
-      child: Padding(
-        padding: const EdgeInsets.only(
-          left: 32.0,
-          top: 16.0,
-          right: 16.0,
-          bottom: 16.0,
-        ),
-        child: Container(
-          child: Text(
-            _orderHeader.purchaseName,
-            style: appThemeData.textTheme.subtitle2!.copyWith(
-              color: appThemeData.colorScheme.onSecondaryContainer,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-
-class OrderHeader {
-  late String purchaseId;
-  late String purchaseName;
-  OrderHeader({
-    required Order order, 
-  }) {
-    purchaseId = '${order['purchase/id']}';
-    purchaseName = '${order['purchase/name']}';
   }
 }
